I expect this to completely implode
