+++ packages/nextjs/src/edge/index.ts
@@ -1,12 +1,6 @@
-import { getCurrentHub, getIntegrationsToSetup, initAndBind, Integrations as CoreIntegrations } from '@sentry/core';
+import { getIntegrationsToSetup, initAndBind, Integrations as CoreIntegrations } from '@sentry/core';
 import type { Options } from '@sentry/types';
-import {
-  createStackParser,
-  GLOBAL_OBJ,
-  logger,
-  nodeStackLineParser,
-  stackParserFromStackParserOptions,
-} from '@sentry/utils';
+import { createStackParser, GLOBAL_OBJ, nodeStackLineParser, stackParserFromStackParserOptions } from '@sentry/utils';
 
 import { getVercelEnv } from '../common/getVercelEnv';
 import { setAsyncLocalStorageAsyncContextStrategy } from './asyncLocalStorageAsyncContextStrategy';
@@ -104,50 +98,10 @@ export function getSentryRelease(fallback?: string): string | undefined {
 }
 
 /**
- * Call `close()` on the current client, if there is one. See {@link Client.close}.
- *
- * @param timeout Maximum time in ms the client should wait to flush its event queue before shutting down. Omitting this
- * parameter will cause the client to wait until all events are sent before disabling itself.
- * @returns A promise which resolves to `true` if the queue successfully drains before the timeout, or `false` if it
- * doesn't (or if there's no client defined).
- */
-export async function close(timeout?: number): Promise<boolean> {
-  const client = getCurrentHub().getClient<EdgeClient>();
-  if (client) {
-    return client.close(timeout);
-  }
-  __DEBUG_BUILD__ && logger.warn('Cannot flush events and disable SDK. No client defined.');
-  return Promise.resolve(false);
-}
-
-/**
- * This is the getter for lastEventId.
- *
- * @returns The last event id of a captured event.
- */
-export function lastEventId(): string | undefined {
-  return getCurrentHub().lastEventId();
-}
-
-/**
  * Just a passthrough in case this is imported from the client.
  */
 export function withSentryConfig<T>(exportedUserNextConfig: T): T {
   return exportedUserNextConfig;
 }
 
-export { flush } from './utils/flush';
-
 export * from '@sentry/core';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryAPI,
-  wrapApiHandlerWithSentry,
-} from './wrapApiHandlerWithSentry';
-
-export { wrapApiHandlerWithSentryVercelCrons } from '../common/wrapApiHandlerWithSentryVercelCrons';
-
-export { wrapMiddlewareWithSentry } from './wrapMiddlewareWithSentry';
-
-export { wrapServerComponentWithSentry } from './wrapServerComponentWithSentry';
