+++ packages/nextjs/src/index.server.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export * from './config';
 export * from './server';
