+++ packages/nextjs/src/index.types.ts
@@ -2,6 +2,7 @@
 
 // We export everything from both the client part of the SDK and from the server part. Some of the exports collide,
 // which is not allowed, unless we redifine the colliding exports in this file - which we do below.
+export * from './common';
 export * from './config';
 export * from './client';
 export * from './server';
@@ -27,9 +28,6 @@ export declare const Integrations: typeof clientSdk.Integrations &
 export declare const defaultIntegrations: Integration[];
 export declare const defaultStackParser: StackParser;
 
-export declare function close(timeout?: number | undefined): PromiseLike<boolean>;
-export declare function flush(timeout?: number | undefined): PromiseLike<boolean>;
-export declare function lastEventId(): string | undefined;
 export declare function getSentryRelease(fallback?: string): string | undefined;
 
 export declare const ErrorBoundary: typeof clientSdk.ErrorBoundary;
