+++ packages/nextjs/src/server/index.ts
@@ -10,7 +10,7 @@ import * as path from 'path';
 import { devErrorSymbolicationEventProcessor } from '../common/devErrorSymbolicationEventProcessor';
 import { getVercelEnv } from '../common/getVercelEnv';
 import { buildMetadata } from '../common/metadata';
-import { isBuild } from './utils/isBuild';
+import { isBuild } from '../common/utils/isBuild';
 
 export * from '@sentry/node';
 export { captureUnderscoreErrorException } from '../common/_error';
@@ -156,49 +156,3 @@ function addServerIntegrations(options: NodeOptions): void {
 const deprecatedIsBuild = (): boolean => isBuild();
 // eslint-disable-next-line deprecation/deprecation
 export { deprecatedIsBuild as isBuild };
-
-export { wrapApiHandlerWithSentryVercelCrons } from '../common/wrapApiHandlerWithSentryVercelCrons';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryGetStaticProps,
-  wrapGetStaticPropsWithSentry,
-} from './wrapGetStaticPropsWithSentry';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryServerSideGetInitialProps,
-  wrapGetInitialPropsWithSentry,
-} from './wrapGetInitialPropsWithSentry';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryServerSideAppGetInitialProps,
-  wrapAppGetInitialPropsWithSentry,
-} from './wrapAppGetInitialPropsWithSentry';
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryServerSideDocumentGetInitialProps,
-  wrapDocumentGetInitialPropsWithSentry,
-} from './wrapDocumentGetInitialPropsWithSentry';
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryServerSideErrorGetInitialProps,
-  wrapErrorGetInitialPropsWithSentry,
-} from './wrapErrorGetInitialPropsWithSentry';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryGetServerSideProps,
-  wrapGetServerSidePropsWithSentry,
-} from './wrapGetServerSidePropsWithSentry';
-
-export {
-  // eslint-disable-next-line deprecation/deprecation
-  withSentry,
-  // eslint-disable-next-line deprecation/deprecation
-  withSentryAPI,
-  wrapApiHandlerWithSentry,
-} from './wrapApiHandlerWithSentry';
-
-export { wrapServerComponentWithSentry } from './wrapServerComponentWithSentry';
